tests/wpt/web-platform-tests/referrer-policy/css-integration/font-face/processing-instruction.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - Font from external stylesheet inserted via a ProcessingInstruction</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="never">
  </head>
  <body>
    <p>Check that resources from external stylesheets (referenced from a
    ProcessingInstruction) are loaded with the referrer and referrer policy
    from the external stylesheet.</p>

    <div class="styled"></div>

    <script>
      let css_test = async_test("Font from external stylesheet (from ProcessingInstruction).");
      let id = token();
      let url_prefix = location.protocol + "//www1." + location.hostname + ":" + location.port;
      let css_url = url_prefix +
                    "/referrer-policy/generic/subresource/stylesheet.py?id=" +
                    id + "&amp;type=font";
      let expected = url_prefix +
                    "/referrer-policy/generic/subresource/stylesheet.py?id=" +
                    id + "&type=font";
      let font_url = url_prefix + "/referrer-policy/generic/subresource/font.py" +
                     "?id=" + id + "&report-headers";

      let processingInstruction =
        document.createProcessingInstruction(
          "xml-stylesheet", "href=\"" + css_url + "\" type=\"text/css\"");
      css_test.step_timeout(
          queryXhr.bind(this, font_url,
                        function(message) {
                          css_test.step(function() {
                            assert_own_property(message, "headers");
                            assert_own_property(message, "referrer");
                            assert_equals(message.referrer, expected);
                          });
                          css_test.done();
                        }),
          1000);
      document.insertBefore(processingInstruction, document.firstChild);
    </script>

    <div id="log"></div>
  </body>
</html>
